Shift classes into its own separate package
