+++ src/main/java/seedu/sherpass/timer/TimerLogic.java
@@ -1,4 +1,4 @@
-package seedu.sherpass.util;
+package seedu.sherpass.timer;
 
 import seedu.sherpass.command.Command;
 import seedu.sherpass.command.MarkCommand;
@@ -8,7 +8,10 @@ import seedu.sherpass.util.parser.TaskParser;
 import seedu.sherpass.util.parser.TimerParser;
 
 import seedu.sherpass.task.TaskList;
+import seedu.sherpass.util.Storage;
+import seedu.sherpass.util.Ui;
 
+import static seedu.sherpass.constant.Index.STUDY_PARAMETER_INDEX;
 import static seedu.sherpass.constant.Message.ERROR_INVALID_TIMER_INPUT_MESSAGE;
 
 public class TimerLogic {
@@ -34,7 +37,7 @@ public class TimerLogic {
      * @param storage Storage.
      * @param parsedInput parsedInput.
      */
-    public void markTask(Storage storage, String[] parsedInput) {
+    public void markTask(Storage storage, String parsedInput) {
         if (isTimerPausedOrStopped()) {
             executeMark(storage, parsedInput);
         } else {
@@ -42,7 +45,8 @@ public class TimerLogic {
         }
     }
 
-    private void executeMark(Storage storage, String[] parsedInput) {
+
+    private void executeMark(Storage storage, String parsedInput) {
         Command c = TaskParser.prepareMarkOrUnmark(parsedInput, MarkCommand.COMMAND_WORD, taskList);
         if (c != null) {
             c.execute(taskList, ui, storage);
@@ -147,12 +151,20 @@ public class TimerLogic {
         return timer.isTimerRunning();
     }
 
+
+    private String selectStudyTimer(String[] parsedInput) {
+        if (parsedInput[STUDY_PARAMETER_INDEX].trim().equals("stopwatch")) {
+            return "stopwatch";
+        }
+        return "countdown";
+    }
+
     /**
      * Resets the timer by creating a new timer object, which can then be started by the user.
      *
      */
     public void callResetTimer(String[] parsedInput) {
-        String parameter = TimerParser.parseStudyParameter(parsedInput);
+        String parameter = selectStudyTimer(parsedInput);
         if (parameter.equals("stopwatch")) {
             timer = new Stopwatch(taskList, ui);
             return;
@@ -160,23 +172,19 @@ public class TimerLogic {
         timer = new Countdown(taskList, ui);
     }
 
+
+
     private boolean isTimerPausedOrStopped() {
         if (!isTimerInitialised) {
             return true;
         }
-        if (timer.isTimerPaused()) {
-            return true;
-        }
-        return false;
+        return timer.isTimerPaused();
     }
 
     public boolean getIsTimerRunning() {
         if (!isTimerInitialised) {
             return false;
         }
-        if (timer.isTimerRunning) {
-            return true;
-        }
-        return false;
+        return timer.isTimerRunning;
     }
 }
